+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ ChangeLog
 - Fix resuming uploads by passing missed ``Upload-Length`` header:
   `#5 <https://github.com/pylotcode/aiohttp-tus/pull/5>`_
 - Add documentation about `CORS Headers <https://aiohttp-tus.readthedocs.io/en/latest/usage.html#cors-headers>`_
+- Allow to provide upload resource name, which can be lately used for URL reversing
 
 1.0.0b2 (2020-03-18)
 ====================